Created a BaseObjects page and imported all common libraries. Imported the base objects to all existing pages.
