+++ robot/EDA/resources/BaseObjects.py
@@ -1,0 +1,27 @@
+from robot.libraries.BuiltIn import BuiltIn
+
+
+class BaseEDAPage:
+    @property
+    def builtin(self):
+        return BuiltIn()
+
+    @property
+    def cumulusci(self):
+        return self.builtin.get_library_instance("cumulusci.robotframework.CumulusCI")
+
+    @property
+    def eda(self):
+        return self.builtin.get_library_instance('EDA')
+
+    @property
+    def pageobjects(self):
+        return self.builtin.get_library_instance("cumulusci.robotframework.PageObjects")
+
+    @property
+    def salesforce(self):
+        return self.builtin.get_library_instance('cumulusci.robotframework.Salesforce')
+
+    @property
+    def selenium(self):
+        return self.builtin.get_library_instance("SeleniumLibrary")
